Kit_of_Masonry + Kit_of_Tablet: page scripts become static assets (QST-0021.4, .5)
- MASONRY_SCRIPT -> app/static/js/masonry.js, served by
  AtlasVenustas/Kit_of_Masonry.masonry_head_tags()
- HOME_SCRIPT -> app/static/js/tablet.js, served by
  AtlasVenustas/Kit_of_Tablet.tablet_head_tags()
- Script content byte-preserved; tags load from /static (already
  mounted via App static_assets), same head positions as before -
  execution order unchanged, now lintable and browser-cacheable
- Both Kits carry __main__ self-tests per Canon
- Masonry's remaining client is .spellcaster-box (NPC sheet left the
  grid in QST-0008); noted in the Kit docstring to retire with
  QST-0026
